# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AudiModal.ai API",
    "description": "Comprehensive AI-powered document processing and analytics platform with ML/AI classification, anomaly detection, unified sync framework, and enterprise integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "AudiModal.ai",
      "url": "https://audimodal.ai",
      "email": "support@audimodal.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.audimodal.ai/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.audimodal.ai/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/ml/insights": {
      "get": {
        "tags": ["ML/AI - Insights"],
        "summary": "List insights",
        "description": "Retrieve generated insights with filtering options",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by insight type",
            "schema": {
              "type": "string",
              "enum": ["content", "usage", "trend", "anomaly", "recommendation", "prediction", "quality", "security", "compliance", "performance"]
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by insight category",
            "schema": {
              "type": "string",
              "enum": ["document_health", "user_engagement", "content_quality", "access_patterns", "storage_optimization", "security_risk", "compliance_gap", "performance_issue"]
            }
          },
          {
            "name": "severity",
            "in": "query",
            "description": "Filter by severity level",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high", "critical"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Insights retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "insights": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DocumentInsight"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "DocumentInsight": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "document_id": {
            "type": "string",
            "format": "uuid"
          },
          "insight_type": {
            "type": "string",
            "enum": ["content", "usage", "trend", "anomaly", "recommendation", "prediction", "quality", "security", "compliance", "performance"]
          },
          "category": {
            "type": "string",
            "enum": ["document_health", "user_engagement", "content_quality", "access_patterns", "storage_optimization", "security_risk", "compliance_gap", "performance_issue"]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": ["low", "medium", "high", "critical"]
          },
          "confidence": {
            "type": "number",
            "format": "float"
          },
          "data": {
            "type": "object"
          },
          "metrics": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "value": {
                  "type": "number",
                  "format": "float"
                },
                "unit": {
                  "type": "string"
                },
                "trend": {
                  "type": "string",
                  "enum": ["up", "down", "stable"]
                }
              }
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "action": {
                  "type": "string"
                },
                "priority": {
                  "type": "string",
                  "enum": ["low", "medium", "high", "urgent"]
                },
                "impact": {
                  "type": "string"
                },
                "effort": {
                  "type": "string"
                },
                "timeline": {
                  "type": "string"
                }
              }
            }
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          },
          "generated_by": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["new", "reviewed", "actioned", "dismissed", "resolved"]
          }
        }
      },
...
    }
  },
...
}