# internal/server/server.go
package server

import (
	"context"
	"fmt"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"syscall"
	"time"

	"github.com/prometheus/client_golang/prometheus/promhttp"

	"github.com/jscharber/audimodal/internal/database"
	"github.com/jscharber/audimodal/internal/server/handlers"
	"github.com/jscharber/audimodal/internal/server/response"
	"github.com/jscharber/audimodal/internal/services"
	"github.com/jscharber/audimodal/pkg/events"
	"github.com/jscharber/audimodal/pkg/health"
	"github.com/jscharber/audimodal/pkg/logger"
	"github.com/jscharber/audimodal/pkg/metrics"
)

// Server represents the HTTP server
type Server struct {
	config          *Config
	db              *database.Database
	httpServer      *http.Server
	router          *Router
	logger          *logger.Logger
	metricsRegistry *metrics.MetricsRegistry
	systemMetrics   *metrics.SystemMetrics
	healthChecker   *health.HealthChecker
	healthHandler   *health.Handler
}

// New creates a new HTTP server
func New(config *Config, db *database.Database) (*Server, error) {
	if err := config.Validate(); err != nil {
		return nil, fmt.Errorf("invalid server config: %w", err)
	}

	// Initialize structured logger
	logLevel := logger.ParseLogLevel(config.LogLevel)
	logFormat := logger.JSONFormat
	if config.LogFormat == "text" {
		logFormat = logger.TextFormat
	}

	serverLogger := logger.NewLogger(&logger.Config{
		Level:        logLevel,
		Format:       logFormat,
		Output:       os.Stdout,
		Service:      "audimodal-server",
		Version:      "1.0.0",
		EnableCaller: config.IsDevelopment(),
		Fields:       make(map[string]interface{}),
	})

	// Initialize metrics registry
	metricsRegistry := metrics.NewRegistry()
	metricsRegistry.SetEnabled(config.MetricsEnabled)
	metricsRegistry.AddGlobalLabel("service", "audimodal")
	metricsRegistry.AddGlobalLabel("version", "1.0.0")

	// Initialize system metrics
	systemMetrics := metrics.NewSystemMetrics(metricsRegistry)

	// Initialize health checker
	healthChecker := health.NewHealthChecker(config.HealthCheckTimeout)

	// Add database health check
	healthChecker.AddChecker(health.DatabaseChecker("database", func(ctx context.Context) error {
		return db.HealthCheck(ctx)
	}))

	// Add system health checks
	healthChecker.AddChecker(health.MemoryChecker("memory", 90.0))
	healthChecker.AddChecker(health.DiskSpaceChecker("disk", "/", 10.0))

	// Create health handler
	healthHandler := health.NewHandler(healthChecker, "audimodal", "1.0.0")

	server := &Server{
		config:          config,
		db:              db,
		logger:          serverLogger,
		metricsRegistry: metricsRegistry,
		systemMetrics:   systemMetrics,
		healthChecker:   healthChecker,
		healthHandler:   healthHandler,
	}

	// Initialize router
	server.router = NewRouter(config, db, serverLogger, metricsRegistry, healthHandler)

	// Create HTTP server
	server.httpServer = &http.Server{
		Addr:         config.GetAddress(),
		Handler:      server.router,
		ReadTimeout:  config.ReadTimeout,
		WriteTimeout: config.WriteTimeout,
		IdleTimeout:  config.IdleTimeout,
	}

	return server, nil
}

// Start starts the HTTP server with graceful shutdown
func (s *Server) Start(ctx context.Context) error {
	// Channel to signal server start errors
	serverErrors := make(chan error, 1)

	// Start system metrics collection
	if s.config.MetricsEnabled {
		go s.systemMetrics.Start(ctx, 15*time.Second)
		s.logger.Info("System metrics collection started")
	}

	// Start server in a goroutine
	go func() {
		s.logger.WithFields(map[string]interface{}{
			"address":     s.config.GetAddress(),
			"tls_enabled": s.config.TLSEnabled,
		}).Info("Starting HTTP server")

		var err error
		if s.config.TLSEnabled {
			s.logger.WithFields(map[string]interface{}{
				"cert_file": s.config.TLSCertFile,
				"key_file":  s.config.TLSKeyFile,
			}).Info("TLS enabled, starting HTTPS server")
			err = s.httpServer.ListenAndServeTLS(s.config.TLSCertFile, s.config.TLSKeyFile)
		} else {
			s.logger.Info("TLS disabled, starting HTTP server")
			err = s.httpServer.ListenAndServe()
		}

		if err != nil && err != http.ErrServerClosed {
			s.logger.WithField("error", err).Error("HTTP server error")
			serverErrors <- err
		}
	}()

	// Wait for server to start or fail
	select {
	case err := <-serverErrors:
		return fmt.Errorf("server failed to start: %w", err)
	case <-time.After(1 * time.Second):
		s.logger.Info("HTTP server started successfully")
	}

	// Set up signal handling for graceful shutdown
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP)

	// Wait for shutdown signal or context cancellation
	select {
	case sig := <-quit:
		s.logger.WithField("signal", sig.String()).Info("Received shutdown signal, initiating graceful shutdown")
	case <-ctx.Done():
		s.logger.Info("Context cancelled, initiating graceful shutdown")
	case err := <-serverErrors:
		s.logger.WithField("error", err).Error("Server error detected, initiating shutdown")
		return err
	}

	return s.Shutdown()
}

// Shutdown gracefully shuts down the server and its dependencies
func (s *Server) Shutdown() error {
	s.logger.Info("Starting graceful shutdown sequence")

	// Create shutdown context with timeout
	ctx, cancel := context.WithTimeout(context.Background(), s.config.ShutdownTimeout)
	defer cancel()

	// Channel to collect shutdown errors
	shutdownErrors := make(chan error, 2)
	var shutdownErr error

	// Start graceful HTTP server shutdown
	go func() {
		s.logger.Info("Shutting down HTTP server")
		if err := s.httpServer.Shutdown(ctx); err != nil {
			shutdownErrors <- fmt.Errorf("HTTP server shutdown error: %w", err)
		} else {
			s.logger.Info("HTTP server stopped gracefully")
			shutdownErrors <- nil
		}
	}()

	// Start database connection cleanup
	go func() {
		s.logger.Info("Closing database connections")
		if err := s.db.Close(); err != nil {
			shutdownErrors <- fmt.Errorf("database shutdown error: %w", err)
		} else {
			s.logger.Info("Database connections closed")
			shutdownErrors <- nil
		}
	}()

	// Stop system metrics collection
	if s.config.MetricsEnabled {
		s.systemMetrics.Stop()
		s.logger.Info("System metrics collection stopped")
	}

	// Wait for both shutdown operations to complete
	for i := 0; i < 2; i++ {
		select {
		case err := <-shutdownErrors:
			if err != nil && shutdownErr == nil {
				shutdownErr = err // Capture first error
				s.logger.WithField("error", err).Error("Shutdown component error")
			}
		case <-ctx.Done():
			if shutdownErr == nil {
				shutdownErr = fmt.Errorf("shutdown timeout exceeded: %v", ctx.Err())
			}
			break
		}
	}

	// Force close if shutdown timeout was exceeded
	if ctx.Err() != nil {
		s.logger.Warn("Graceful shutdown timeout exceeded, forcing server close")
		if err := s.httpServer.Close(); err != nil {
			s.logger.WithField("error", err).Error("Force close error")
		}
	}

	if shutdownErr != nil {
		s.logger.WithField("error", shutdownErr).Error("Shutdown completed with errors")
		return shutdownErr
	}

	s.logger.Info("Graceful shutdown completed successfully")
	return nil
}

// Router represents the HTTP router
type Router struct {
	*http.ServeMux
	config          *Config
	db              *database.Database
	logger          *logger.Logger
	metricsRegistry *metrics.MetricsRegistry
	healthHandler   *health.Handler
	middleware      *MiddlewareStack
}

// NewRouter creates a new HTTP router
func NewRouter(config *Config, db *database.Database, logger *logger.Logger, metricsRegistry *metrics.MetricsRegistry, healthHandler *health.Handler) *Router {
	router := &Router{
		ServeMux:        http.NewServeMux(),
		config:          config,
		db:              db,
		logger:          logger,
		metricsRegistry: metricsRegistry,
		healthHandler:   healthHandler,
		middleware:      NewMiddlewareStack(),
	}

	router.setupMiddleware()
	router.setupRoutes()

	return router
}

// ServeHTTP implements the http.Handler interface
func (r *Router) ServeHTTP(w http.ResponseWriter, req *http.Request) {
	// Apply middleware stack
	handler := r.middleware.Apply(r.ServeMux)
	handler.ServeHTTP(w, req)
}

// setupMiddleware configures the middleware stack
func (r *Router) setupMiddleware() {
	// Add middleware in order of execution
	r.middleware.Use(RecoveryMiddleware())
	r.middleware.Use(SecurityHeadersMiddleware())
	r.middleware.Use(RequestIDMiddleware(r.config.RequestIDHeader))

	// Add metrics collection middleware if enabled
	if r.config.MetricsEnabled {
		r.middleware.Use(metrics.GetHTTPMetricsMiddleware())
	}

	// Add structured logging middleware if request logging is enabled
	if r.config.LogRequests {
		httpLogConfig := &logger.HTTPLogConfig{
			SkipPaths:       []string{r.config.HealthCheckPath, r.config.MetricsPath},
			LogRequestBody:  r.config.LogRequestBody,
			LogResponseBody: r.config.LogResponseBody,
			LogHeaders:      r.config.LogHeaders,
			SanitizeHeaders: []string{
				"authorization", "x-api-key", "cookie", "set-cookie",
				"x-auth-token", "x-csrf-token", "jwt", "bearer",
			},
			MaxBodySize: 1024, // 1KB
		}
		r.middleware.Use(logger.RequestLoggingMiddleware(r.logger, httpLogConfig))
	}

	r.middleware.Use(CORSMiddleware(r.config))
	r.middleware.Use(RateLimitMiddleware(r.config))
	r.middleware.Use(MaxRequestSizeMiddleware(r.config.MaxRequestSize))
	r.middleware.Use(ContentTypeMiddleware())
	r.middleware.Use(PaginationMiddleware(r.config))
}

// setupRoutes configures all API routes
func (r *Router) setupRoutes() {
	r.logger.Info("Setting up routes", "api_prefix", r.config.APIPrefix)

	// Health check endpoints (no auth required) - Register these FIRST to avoid conflicts
	r.ServeMux.HandleFunc(r.config.HealthCheckPath, r.healthHandler.HealthCheckHandler())
	r.ServeMux.HandleFunc(r.config.HealthCheckPath+"/ready", r.healthHandler.ReadinessHandler())
	r.ServeMux.HandleFunc(r.config.HealthCheckPath+"/live", r.healthHandler.LivenessHandler())

	// Metrics endpoints (no auth required)
	if r.config.MetricsEnabled {
		r.ServeMux.HandleFunc(r.config.MetricsPath, r.metricsHandler)
		// Add Prometheus-formatted metrics endpoint
		r.ServeMux.Handle("/metrics/prometheus", promhttp.Handler())
	}

	// API routes with authentication
	apiPrefix := r.config.APIPrefix

	// Create storage service first (needed by file handler)
	encryptionKey := os.Getenv("EAI_ENCRYPTION_KEY")
	if encryptionKey == "" {
		encryptionKey = "your-32-byte-encryption-key-here"
	}
	storageService := services.NewStorageService(r.db, []byte(encryptionKey))
	storageHandler := handlers.NewStorageHandler(r.db, storageService)

	// Initialize Kafka event producer if enabled (using pure Go kafka-go library)
	var eventProducer *events.SimpleProducer
	if r.config.Kafka != nil && r.config.Kafka.Enabled {
		producerConfig := events.SimpleProducerConfig{
			BootstrapServers: r.config.Kafka.BootstrapServers,
			ClientID:         r.config.Kafka.ClientID,
		}
		var err error
		eventProducer, err = events.NewSimpleProducer(producerConfig)
		if err != nil {
			r.logger.Warn("Failed to initialize Kafka producer, events will not be published", "error", err)
		} else {
			r.logger.Info("Kafka event producer initialized", "brokers", r.config.Kafka.BootstrapServers)
		}
	}

	// Create handlers
	tenantHandler := handlers.NewTenantHandler(r.db)
	dataSourceHandler := handlers.NewDataSourceHandler(r.db)
	sessionHandler := handlers.NewProcessingSessionHandler(r.db)
	dlpHandler := handlers.NewDLPHandler(r.db)
	fileHandler := handlers.NewFileHandler(r.db, storageService, eventProducer)
	chunkHandler := handlers.NewChunkHandler(r.db)
	webHandler := handlers.NewWebHandler(r.db, r.logger)
	mlAnalysisHandler := handlers.NewMLAnalysisHandler(r.db, r.logger)

	// Apply authentication middleware to API routes
	authMiddleware := AuthenticationMiddleware(r.config, r.db)

	// Authentication bypass for development - configure properly in production
	noAuthMiddleware := func(h http.Handler) http.Handler { return h }

	// Tenant management routes
	tenantsPath := fmt.Sprintf("%s/tenants", apiPrefix)
	r.logger.Info("Registering tenant management route", "path", tenantsPath)
	r.ServeMux.Handle(tenantsPath, noAuthMiddleware(http.HandlerFunc(tenantHandler.HandleTenant)))

	// Tenant-scoped routes (require tenant context)
	tenantMiddleware := TenantMiddleware(r.db)
	tenantAuth := func(h http.Handler) http.Handler {
		return authMiddleware(tenantMiddleware(h))
	}

	// All tenant-scoped routes - try Go 1.22+ wildcard pattern
	tenantsWildcard := fmt.Sprintf("%s/tenants/{id...}", apiPrefix)
	r.logger.Info("Registering tenant wildcard route", "path", tenantsWildcard)
	r.ServeMux.Handle(tenantsWildcard, noAuthMiddleware(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
		// Check if this is a tenant management operation (e.g., GET /api/v1/tenants/{id})
		pathParts := strings.Split(strings.TrimPrefix(req.URL.Path, "/"), "/")
		if len(pathParts) == 4 && pathParts[0] == "api" && pathParts[1] == "v1" && pathParts[2] == "tenants" {
			// This is a tenant management operation
			// Routing to tenant handler
			tenantHandler.HandleTenant(w, req)
			return
		}

		// Otherwise, it's a tenant-scoped operation that requires tenant context
		// Log tenant-scoped route access
		r.logger.Info("Tenant-scoped route called",
			"path", req.URL.Path,
			"method", req.Method,
			"request_id", getRequestID(req.Context()))

		// Apply tenant middleware and dispatch
		tenantAuth(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
			r.logger.Info("After tenant middleware",
				"path", req.URL.Path,
				"tenant_context", getTenantContext(req.Context()))

			// Dispatch to appropriate handler based on path
			if isDataSourceRoute(req.URL.Path, apiPrefix) {
				// Route to data source handler
				dataSourceHandler.ServeHTTP(w, req)
			} else if isSessionRoute(req.URL.Path, apiPrefix) {
				// Route to session handler
				sessionHandler.ServeHTTP(w, req)
			} else if isDLPRoute(req.URL.Path, apiPrefix) {
				// Route to DLP handler
				dlpHandler.ServeHTTP(w, req)
			} else if isStorageRoute(req.URL.Path, apiPrefix) {
				// Route to storage handler
				storageHandler.ServeHTTP(w, req)
			} else if isFileRoute(req.URL.Path, apiPrefix) {
				r.logger.Info("Routing to file handler", "path", req.URL.Path)
				fileHandler.ServeHTTP(w, req)
			} else if isChunkRoute(req.URL.Path, apiPrefix) {
				// Route to chunk handler
				chunkHandler.ServeHTTP(w, req)
			} else if isMLAnalysisRoute(req.URL.Path, apiPrefix) {
				// Route to ML analysis handler
				mlAnalysisHandler.ServeHTTP(w, req)
			} else {
				r.logger.Warn("No matching route found", "path", req.URL.Path)
				http.NotFound(w, req)
			}
		})).ServeHTTP(w, req)
	})))

	// Embedding API routes (must be registered before the catch-all apiRootHandler)
	embeddingHandler, err := handlers.NewEmbeddingHandler()
	if err != nil {
		r.logger.Warn("Failed to initialize embedding handler, embedding routes will be disabled", "error", err.Error())
	} else {
		r.logger.Info("Registering embedding API routes")
		// Register embedding routes with explicit paths
		// Note: Only registering routes that don't require path variables (mux.Vars)
		// Routes with path variables need adapter methods to use r.PathValue() instead
		embeddingsPrefix := fmt.Sprintf("%s/embeddings", apiPrefix)

		// Document processing (POST doesn't need path vars)
		r.ServeMux.HandleFunc(fmt.Sprintf("POST %s/documents", embeddingsPrefix), embeddingHandler.ProcessDocument)

		// Search
		r.ServeMux.HandleFunc(fmt.Sprintf("POST %s/search", embeddingsPrefix), embeddingHandler.SearchDocuments)

		// Datasets (GET list and POST create don't need path vars)
		r.ServeMux.HandleFunc(fmt.Sprintf("GET %s/datasets", embeddingsPrefix), embeddingHandler.ListDatasets)
		r.ServeMux.HandleFunc(fmt.Sprintf("POST %s/datasets", embeddingsPrefix), embeddingHandler.CreateDataset)

		// Health and stats
		r.ServeMux.HandleFunc(fmt.Sprintf("GET %s/health", embeddingsPrefix), embeddingHandler.HealthCheck)
		r.ServeMux.HandleFunc(fmt.Sprintf("GET %s/stats", embeddingsPrefix), embeddingHandler.GetServiceStats)
	}

	// API documentation route
	r.logger.Info("Registering API docs and root", "docs_path", fmt.Sprintf("%s/docs", apiPrefix), "root_path", fmt.Sprintf("%s/", apiPrefix))
	r.ServeMux.HandleFunc(fmt.Sprintf("%s/docs", apiPrefix), r.docsHandler)
	r.ServeMux.HandleFunc(fmt.Sprintf("%s/", apiPrefix), r.apiRootHandler)

	// Web UI routes
	r.ServeMux.HandleFunc("/", webHandler.RedirectToLogin())
	r.ServeMux.HandleFunc("/login", webHandler.LoginHandler())
	r.ServeMux.HandleFunc("/dashboard", webHandler.DashboardHandler())
	r.ServeMux.HandleFunc("/static/", webHandler.StaticFileHandler())
}

// metricsHandler handles metrics requests
func (r *Router) metricsHandler(w http.ResponseWriter, req *http.Request) {
	if req.Method != http.MethodGet {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	// Use the metrics registry handler
	if r.metricsRegistry != nil {
		handler := metrics.HTTPMetricsHandler(r.metricsRegistry)
		handler.ServeHTTP(w, req)
		return
	}

	// Fallback to basic response if metrics are disabled
	http.Error(w, "Metrics collection is disabled", http.StatusServiceUnavailable)
}

// docsHandler handles API documentation requests
func (r *Router) docsHandler(w http.ResponseWriter, req *http.Request) {
	if req.Method != http.MethodGet {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	docs := map[string]interface{}{
		"title":       "eAIIngest API",
		"version":     "1.0.0",
		"description": "Enterprise AI Document Processing Platform API",
		"base_url":    fmt.Sprintf("http://%s%s", req.Host, r.config.APIPrefix),
		"endpoints": map[string]interface{}{
			"tenants":             fmt.Sprintf("%s/tenants", r.config.APIPrefix),
			"data_sources":        fmt.Sprintf("%s/tenants/{tenant_id}/data-sources", r.config.APIPrefix),
			"processing_sessions": fmt.Sprintf("%s/tenants/{tenant_id}/sessions", r.config.APIPrefix),
			"dlp_policies":        fmt.Sprintf("%s/tenants/{tenant_id}/dlp-policies", r.config.APIPrefix),
			"files":               fmt.Sprintf("%s/tenants/{tenant_id}/files", r.config.APIPrefix),
			"chunks":              fmt.Sprintf("%s/tenants/{tenant_id}/chunks", r.config.APIPrefix),
			"storage":             fmt.Sprintf("%s/tenants/{tenant_id}/storage", r.config.APIPrefix),
		},
		"authentication": map[string]interface{}{
			"type":        "API Key or JWT Bearer Token",
			"header":      r.config.APIKeyHeader,
			"bearer_auth": "Authorization: Bearer <token>",
		},
	}

	requestID := getRequestID(req.Context())
	response.WriteSuccess(w, requestID, docs, nil)
}

// apiRootHandler handles API root requests
func (r *Router) apiRootHandler(w http.ResponseWriter, req *http.Request) {
	if req.Method != http.MethodGet {
		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
		return
	}

	info := map[string]interface{}{
		"name":        "eAIIngest API",
		"version":     "1.0.0",
		"description": "Enterprise AI Document Processing Platform",
		"status":      "active",
		"endpoints": map[string]string{
			"health":        r.config.HealthCheckPath,
			"metrics":       r.config.MetricsPath,
			"documentation": fmt.Sprintf("%s/docs", r.config.APIPrefix),
		},
	}

	requestID := getRequestID(req.Context())
	response.WriteSuccess(w, requestID, info, nil)
}

// Route helper functions
func isDataSourceRoute(path, apiPrefix string) bool {
	return contains(path, "/data-sources")
}

func isSessionRoute(path, apiPrefix string) bool {
	return contains(path, "/sessions")
}

func isDLPRoute(path, apiPrefix string) bool {
	return contains(path, "/dlp-policies") || contains(path, "/violations")
}

func isFileRoute(path, apiPrefix string) bool {
	return strings.Contains(path, "/files")
}

func isChunkRoute(path, apiPrefix string) bool {
	return contains(path, "/chunks")
}

func isStorageRoute(path, apiPrefix string) bool {
	return contains(path, "/storage")
}

func isMLAnalysisRoute(path, apiPrefix string) bool {
	return strings.Contains(path, "/ml-analysis") || strings.Contains(path, "/ml/insights")
}

func contains(s, substr string) bool {
	result := len(s) >= len(substr) && s[len(s)-len(substr):] == substr ||
		len(s) > len(substr) && s[len(s)-len(substr)-1:len(s)-len(substr)] == "/" && s[len(s)-len(substr):] == substr ||
		s == substr ||
		(len(s) > len(substr) && s[:len(substr)] == substr && s[len(substr)] == '/')

	// Log routing decisions - using proper logger would go here
	return result
}

// RunServer is a convenience function to run the server
func RunServer(config *Config, db *database.Database) error {
	server, err := New(config, db)
	if err != nil {
		return fmt.Errorf("failed to create server: %w", err)
	}

	ctx := context.Background()
	return server.Start(ctx)
}